+++ example/lib/main_space.dart
@@ -42,6 +42,5 @@ class MyGame extends Game {
 
   @override
   Color backgroundColor() => const Color(0xFF3f3f74);
-
 }
