+++ example/pubspec.yaml
@@ -1,5 +1,5 @@
-name: example
-description: flame scrolling sprite example
+name: flame_scrolling_sprite_example
+description: a flame scrolling sprite example
 
 version: 1.0.0+1
